+++ src/shared/hooks/point/index.ts
@@ -1,3 +1,5 @@
-export { useGetPointInfo, useChargePoint, useWithdrawPoint } from './usePoint';
+export { useGetPointInfo } from './useGetPointInfo';
 export { useGetPointLogs } from './useGetPointLogs';
 export { getPointStatusLabel } from './pointStatusMapping';
+export { useChargePoint } from './useChargePoint';
+export { useWithdrawPoint } from './useWithdrawPoint';
