+++ src/shared/hooks/point/useGetPointInfo.ts
@@ -1,0 +1,14 @@
+import {
+  getPointInfo,
+  getPointInfoQueryKey,
+  PointData,
+} from '../../api/point/point.api';
+import { useQuery } from '@tanstack/react-query';
+
+// 포인트 조회
+export const useGetPointInfo = () => {
+  return useQuery<PointData, Error>({
+    queryKey: getPointInfoQueryKey,
+    queryFn: getPointInfo,
+  });
+};
